image-service/requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.33.2
accelerate==0.22.0
safetensors==0.3.3
Pillow==10.0.0
xformers==0.0.21